Move `rdp_opt_set()` and `rdp_opt_reset()` to `csh_defaults.c`
These were previously in src/vmem_client_slash_ftp.c
